(save) Rethinking the database model
OK! A lot fo things happened during this commit.
Since I'm just learning couchdb my opinion of how
couch should be used is ever changing.

In this commit I went back to having a channel-db,
instead of a db pr channel.

I also understood that couch stores revisions of
changed documents. So if I use a predictable id that
will be globally unique Im able to easily update and
keep revisions of documents inside couch. Thus, we
do not need the 'latest'-method anymore for trying
to fetch the latest revision of a document. Wee!