(refac) Added Middleware for inbound messages
May reconsider the interface for adding in- and outbound
middleware. The current interface is useIn and useOut
respectively.

Next, I do want to specify what irc "container" to use
in the middleware. As it is specified on proxy init time.
Somehow I do want the "containers" to be added to the
request and used later on, if no container is found we can
default the one we are currently using.